+++ lib/mix/tasks/phoenix_test_playwright/test.websocket.ex
@@ -49,7 +49,7 @@ defmodule Mix.Tasks.PhoenixTestPlaywright.Test.Websocket do
       :phoenix_test
       |> Application.get_env(:playwright, [])
       |> Keyword.put(:ws_endpoint, ws_endpoint)
-      |> Keyword.put(:browser_pool, nil)
+      |> Keyword.put(:browser_pool, false)
 
     Application.put_env(:phoenix_test, :playwright, playwright_config)