+++ test/phoenix_test/playwright/no_browser_pool_test.exs
@@ -1,5 +1,5 @@
 defmodule PhoenixTest.Playwright.NoBrowserPoolTest do
-  use PhoenixTest.Playwright.Case, async: true, browser_pool: nil
+  use PhoenixTest.Playwright.Case, async: true, browser_pool: false
 
   test "launches new browser instead of checking out from pool", %{conn: conn} do
     conn