+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 <!-- and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). -->
 
 ## [Unreleased]
-### Fixed
-- Recommend `browser_pool: nil` when using `ws_endpoint` (remote server has a single pre-launched browser).
+### Changed
+- Use `browser_pool: false` (instead of `nil`) to disable browser pooling.
+- Recommend `browser_pool: false` when using `ws_endpoint` (remote server has a single pre-launched browser).
 
 ## [0.12.0] 2026-02-22
 ### Added